Refactoring: Moved IIconLoader into gui-module; some little Refactorings
