+++ gui/src/main/java/de/adito/git/gui/actions/CherryPickAction.java
@@ -7,7 +7,7 @@ import de.adito.git.api.IRepository;
 import de.adito.git.api.ISaveUtil;
 import de.adito.git.api.data.*;
 import de.adito.git.api.exception.AditoGitException;
-import de.adito.git.api.icon.IIconLoader;
+import de.adito.git.gui.icon.IIconLoader;
 import de.adito.git.api.prefs.IPrefStore;
 import de.adito.git.api.progress.IAsyncProgressFacade;
 import de.adito.git.gui.Constants;
@@ -40,9 +40,9 @@ class CherryPickAction extends AbstractTableAction
   private final ISaveUtil saveUtil;
 
   @Inject
-  CherryPickAction(IPrefStore pPrefStore, IIconLoader pIconLoader, INotifyUtil pNotifyUtil, IDialogProvider pDialogProvider, IAsyncProgressFacade pProgressFacade, ISaveUtil pSaveUtil,
-                   IActionProvider pActionProvider, @Assisted Observable<Optional<IRepository>> pRepository,
-                   @Assisted Observable<Optional<List<ICommit>>> pSelectedCommits)
+  CherryPickAction(IPrefStore pPrefStore, IIconLoader pIconLoader, INotifyUtil pNotifyUtil, IDialogProvider pDialogProvider,
+                   IAsyncProgressFacade pProgressFacade, ISaveUtil pSaveUtil, IActionProvider pActionProvider,
+                   @Assisted Observable<Optional<IRepository>> pRepository, @Assisted Observable<Optional<List<ICommit>>> pSelectedCommits)
   {
     super(ACTION_NAME, _getIsEnabledObservable(pSelectedCommits));
     saveUtil = pSaveUtil;
