+++ gui/src/main/java/de/adito/git/gui/actions/CommitAction.java
@@ -6,7 +6,7 @@ import de.adito.git.api.IRepository;
 import de.adito.git.api.ISaveUtil;
 import de.adito.git.api.data.IConfig;
 import de.adito.git.api.data.IFileChangeType;
-import de.adito.git.api.icon.IIconLoader;
+import de.adito.git.gui.icon.IIconLoader;
 import de.adito.git.api.prefs.IPrefStore;
 import de.adito.git.api.progress.IAsyncProgressFacade;
 import de.adito.git.gui.Constants;
