+++ gui/src/main/java/de/adito/git/gui/actions/ExpandTreeAction.java
@@ -2,7 +2,7 @@ package de.adito.git.gui.actions;
 
 import com.google.inject.Inject;
 import com.google.inject.assistedinject.Assisted;
-import de.adito.git.api.icon.IIconLoader;
+import de.adito.git.gui.icon.IIconLoader;
 import de.adito.git.gui.Constants;
 
 import javax.swing.*;
