+++ gui/src/main/java/de/adito/git/gui/actions/FileHistoryAction.java
@@ -3,7 +3,7 @@ package de.adito.git.gui.actions;
 import com.google.inject.Inject;
 import com.google.inject.assistedinject.Assisted;
 import de.adito.git.api.IRepository;
-import de.adito.git.api.icon.IIconLoader;
+import de.adito.git.gui.icon.IIconLoader;
 import de.adito.git.gui.Constants;
 import de.adito.git.gui.window.IWindowProvider;
 import de.adito.git.impl.data.CommitFilterImpl;
