+++ gui/src/main/java/de/adito/git/gui/dialogs/CommitDialog.java
@@ -7,7 +7,7 @@ import com.jidesoft.swing.CheckBoxTreeSelectionModel;
 import de.adito.git.api.*;
 import de.adito.git.api.data.IFileChangeType;
 import de.adito.git.api.data.IFileStatus;
-import de.adito.git.api.icon.IIconLoader;
+import de.adito.git.gui.icon.IIconLoader;
 import de.adito.git.api.prefs.IPrefStore;
 import de.adito.git.gui.Constants;
 import de.adito.git.gui.IEditorKitProvider;
@@ -72,9 +72,10 @@ class CommitDialog extends AditoBaseDialog<CommitDialogResult> implements IDisca
   private final Disposable disposable;
 
   @Inject
-  public CommitDialog(IFileSystemUtil pFileSystemUtil, IQuickSearchProvider pQuickSearchProvider, IActionProvider pActionProvider, IIconLoader pIconLoader,
-                      IPrefStore pPrefStore, @Assisted IDialogDisplayer.IDescriptor pIsValidDescriptor, @Assisted Observable<Optional<IRepository>> pRepository,
-                      @Assisted Observable<Optional<List<IFileChangeType>>> pFilesToCommit, @Assisted String pMessageTemplate, IEditorKitProvider pEditorKitProvider)
+  public CommitDialog(IFileSystemUtil pFileSystemUtil, IQuickSearchProvider pQuickSearchProvider, IActionProvider pActionProvider,
+                      IIconLoader pIconLoader, IPrefStore pPrefStore, @Assisted IDialogDisplayer.IDescriptor pIsValidDescriptor,
+                      @Assisted Observable<Optional<IRepository>> pRepository, @Assisted Observable<Optional<List<IFileChangeType>>> pFilesToCommit,
+                      @Assisted String pMessageTemplate, IEditorKitProvider pEditorKitProvider)
   {
     actionProvider = pActionProvider;
     iconLoader = pIconLoader;
