+++ gui/src/main/java/de/adito/git/gui/dialogs/RevertFilesDialog.java
@@ -3,7 +3,7 @@ package de.adito.git.gui.dialogs;
 import com.google.inject.assistedinject.Assisted;
 import de.adito.git.api.IFileSystemUtil;
 import de.adito.git.api.data.IFileChangeType;
-import de.adito.git.api.icon.IIconLoader;
+import de.adito.git.gui.icon.IIconLoader;
 import de.adito.git.api.prefs.IPrefStore;
 import de.adito.git.gui.Constants;
 import de.adito.git.gui.actions.IActionProvider;
