+++ gui/src/main/java/de/adito/git/gui/dialogs/panels/basediffpanel/MergePanel.java
@@ -3,7 +3,7 @@ package de.adito.git.gui.dialogs.panels.basediffpanel;
 import de.adito.git.api.IDiscardable;
 import de.adito.git.api.data.EChangeSide;
 import de.adito.git.api.data.IMergeDiff;
-import de.adito.git.api.icon.IIconLoader;
+import de.adito.git.gui.icon.IIconLoader;
 import de.adito.git.gui.Constants;
 import de.adito.git.gui.IEditorKitProvider;
 import de.adito.git.gui.dialogs.panels.basediffpanel.diffpane.LineNumbersColorModel;
