+++ gui/src/main/java/de/adito/git/gui/icon/SwingIconLoaderImpl.java
@@ -1,6 +1,5 @@
 package de.adito.git.gui.icon;
 
-import de.adito.git.api.icon.IIconLoader;
 import org.jetbrains.annotations.*;
 
 import javax.swing.*;
@@ -15,7 +14,7 @@ public class SwingIconLoaderImpl implements IIconLoader
    * {@inheritDoc}
    */
   @Override
-  public ImageIcon getIcon(@NotNull String pIconBase)
+  public ImageIcon getIcon(String pIconBase)
   {
     return new ImageIcon(getClass().getResource(pIconBase));
   }
